Moving all Files according to module
