+++ lib/bambora/client.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-require 'bambora/rest_client'
-require 'bambora/headers'
-require 'bambora/json_client'
-require 'bambora/v1/batch_payment_report_resource'
-require 'bambora/v1/batch_payment_resource'
-require 'bambora/v1/profile_resource'
-require 'bambora/v1/payment_resource'
-require 'bambora/version'
+require 'bambora/client/rest_client'
+require 'bambora/client/headers'
+require 'bambora/client/json_client'
+require 'bambora/client/v1/batch_payment_report_resource'
+require 'bambora/client/v1/batch_payment_resource'
+require 'bambora/client/v1/profile_resource'
+require 'bambora/client/v1/payment_resource'
+require 'bambora/client/version'
 
 module Bambora
   module Client
