+++ lib/bambora/client/version.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
 module Bambora
-  VERSION = '0.1.0'
+  module Client
+    VERSION = '0.1.0'
+  end
 end
