--- spec/bambora_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe Bambora do
-  it 'has a version number' do
-    expect(Bambora::VERSION).not_to be nil
-  end
-end
